Create 3Sum_1
Sort of nums[] must be performed and it's not the key influence for the efficiency.
Almost all solutions are based on twice nesting circulations. Using the value of the first number we can simplify the 3-Sum problem to a find-one-number problem.
However, the usage of malloc() and calloc() is still a little bit confusing.