Generate Parentheses according to given brackets couple number
Using the two way recursion to get possible next '(' or ')' according to the number of left symbol and right symbol. Each choic of left and right would create a new branch. Only when the last position if filled, the size could be added up. The change of () is come from back to front, as can be seen from monitor print result.